+++ lib/light-service/organizer/with_reducer_log_decorator.rb
@@ -5,10 +5,13 @@ module LightService
 
       alias logged? logged
 
-      def initialize(organizer, decorated = WithReducer.new)
+      def initialize(organizer, decorated: WithReducer.new, logger:)
         @decorated = decorated
         @organizer = organizer
-        @logger = LightService::Configuration.logger
+
+        decorated.organizer = organizer
+
+        @logger = logger
         @logged = false
       end