+++ spec/acceptance/organizer/reduce_until_spec.rb
@@ -40,4 +40,10 @@ RSpec.describe LightService::Organizer do
     result = TestReduceUntil.call(empty_context)
     expect(result).to be_success
   end
+
+  it "is expected to know its organizer when reducing until a condition" do
+    result = TestReduceUntil.call(:number => 1)
+
+    expect(result.organized_by).to eq TestReduceUntil
+  end
 end