+++ spec/organizer_spec.rb
@@ -19,6 +19,11 @@ describe LightService::Organizer do
       result = TestDoubles::AnOrganizer.call(:user => user)
       expect(result).to eq(ctx)
     end
+
+    it "sets itself as the organizer" do
+      result = TestDoubles::AnOrganizer.call(:user => user)
+      expect(result.organized_by).to eq TestDoubles::AnOrganizer
+    end
   end
 
   context "when #with is called with Context" do
@@ -88,6 +93,22 @@ describe LightService::Organizer do
 
     it "returns the context" do
       expect(reduced).to eq(ctx)
+    end
+  end
+
+  context 'can add items to the context' do
+    specify 'with #add_to_context' do
+      result = TestDoubles::AnOrganizerThatAddsToContext.call
+      expect(result[:strongest_avenger]).to eq 'The Thor'
+      expect(result[:last_jedi]).to eq 'Rey'
+    end
+  end
+
+  context 'can assign key aliaeses' do
+    it 'with #add_aliases' do
+      result = TestDoubles::AnOrganizerThatAddsAliases.call
+      expect(result[:foo]).to eq :bar
+      expect(result[:baz]).to eq :bar
     end
   end
 end