+++ spec/spec_helper.rb
@@ -15,8 +15,9 @@ end
 require 'light-service'
 require 'light-service/testing'
 require 'ostruct'
-require 'active_support/core_ext/string'
 require 'pry'
 require 'support'
+require 'test_doubles'
+require 'stringio'
 
 I18n.enforce_available_locales = true